Generalize the signature of `toBStream`.
I'm working on an app that uses a third-party library operating on
`ConduitT () ByteString m ()` values. I can't pass that in to the
`toBStream` function that this library provides, as the input type is
rigid and hidden behind the `Producer`'s `forall`. I was able to
generalize it properly by copying `toBStream` into my project and
generalizing the `Producer` to a `ConduitT ()`, and figured that this
might prove useful upstream.

Because every `Producer` can be generalized to the above `ConduitT`
type, I haven't elected to add a `toBStreamProducer` function. GHC has
informed me that all uses of `Conduit`, `Producer`, and `Consumer` are
deprecated; I'd be happy to generalize those to equivalent
`ConduitT`s, but elected to keep the surface area of this patch as
small as possible.